backend/requirements.txt
fastapi==0.115.4
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
starlette==0.41.2
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
uvicorn==0.32.0
pytest==8.3.4
python-multipart==0.0.19
httpx==0.28.1